requirements.txt
Jinja2==2.7.2
requests==2.2.1
lxml>=2.3.2
boto==2.27.0
pyOpenSSL>=16.2.0

